requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Google AI / ADK
google-genai>=1.0.0
google-adk>=0.3.0
google-cloud-aiplatform>=1.87.0

# Google API Integrations (Calendar, Gmail, Tasks)
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0

# Database
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.32
alembic>=1.13.0
pgvector>=0.3.3

# Scheduler
apscheduler>=3.10.0

# Workflow intelligence + retries
networkx>=3.3
tenacity>=8.3.0
imap-tools>=1.7.2
logfire>=0.46.1

# HTTP client
httpx>=0.27.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.0.0
python-dateutil>=2.9.0
